Improve JIT support
Modify Makefile to configure and build JitBuilder library from the
Eclipse OMR project. Also add targets to build the lpjit.so library
from rosie-lpeg.

Change rosie-run to automatically add lib diretory to LD_LIBRARY_PATH
so that lpjit.so can be loaded on demand.

Modify run.lua to detect three new new command-line options related
to JIT compilation:
   -jit causes the JIT to be loaded.
   -jitsizelimit N  causes only patterns with fewer than N bytecodes
                    to be JIT compiled (N=3000 by default)
   -jitmatchcount N causes a pattern to be JIT compiled after
                    match is called on it N times (N=1 by default)
                    unless the pattern has too many bytecodes

Also change the timing of the call to cleanup() so the JIT will only
be unloaded at the very end of execution.

Finally, add a new target "testwithjit" that does the same testing
as "make test" but sets -jit -jitmatchcount 0 so that as many
patterns as possible are JIT compiled before calling match. This
make target should be used to test any changes that might affect
the JIT.

Signed-off-by: Mark Stoodley <mstoodle@ca.ibm.com>